Use rpc.Connection instead of rpc.GenericClient
